Rename MyObjectMixin a better name for VirtualPurchase views and move to mixins.py. Improve virtual purchase deletion template. Fix bug in momentum plot if 200 datapoints not available
